requirements.txt
bcrypt==3.2.2
brotlipy==0.7.0
certifi==2021.10.8
chardet==5.0.0
click==8.1.3
distlib==0.3.5
filelock==3.7.1
Flask==2.2.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.1
numpy==1.23.1
Pillow==9.2.0
platformdirs==2.5.2
pycosat==0.6.3
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
SQLAlchemy==1.4.39
texttable==1.6.4
virtualenv==20.16.3
Werkzeug==2.2.1
wincertstore==0.2
WTForms==3.0.1
xlwt==1.3.0
zipp==3.8.1
